fix(test): TradeControllerTest.testCreateTradeValidationFailure_MissingBook - Test is now getting the correct status code, and testing for the correct body string.
Problem: testCreateTradeValidationFailure_MissingBook() was expecting a 400 status code but was getting 201.
Root Cause: There were no checks in the tradeController or tradeService that would throw an exception if the book (or counterparty) was missing. And in the tradeControllerTest, there were no statements to stub the behaviour of those checks.
Solution: Added statements to throw new RuntimeException with the message ‘Book and Counterparty are required’ in TradeService.populateReferenceDataByName() if the book or counterparty’s names and ids are null. Added a stubbing statement in TradeControllerTest to return the RuntimeException when populateReferenceDataByName is called in the test. Changed the test’s expected error message to ‘Error creating trade: Book and Counterparty are required’ in order to match the ResponseEntity body syntax returned by the TradeController.createTrade().
Impact: The test is now receiving the 400 status code, and is expecting the correct string content. The method tradeService.saveTrade is never called because the checks for book ID and Name are executed beforehand.