fix(test): TradeControllerTest.testUpdateTradeIdMismatch - Added an if statement to check for mismatching IDs, and to throw an exception if not matching. Changed expected string in test.
Problem: testUpdateTradeIdMismatch() is getting a status code of 200 instead of 400.
Root Cause: In TradeController.updateTrade(), the statement tradeDTO.setTradeId(id) immediately sets the given tradeDTO’s ID to the given ID. This means that when the method is given parameters with mismatching IDs, the DTO is immediately overwritten, so the mismatch no longer exists.
Solution: Added an if statement in TradeController.updateTrade() that returns a RuntimeException if there are mismatching IDs. Changed the test’s expected string to ‘Error updating trade: Trade ID in path must match Trade ID in request body’ to match the syntax returned from TradeController.updateTrade.java
Impact: The test is now passing. The tradeDTO ID is no longer getting overwritten before a check for mismatching IDs can be done. And the test is now checking for the correct string in the ResponseEntity body.