fix(test): BookServiceTest.testFindBookByNonExistentId - Added a Mock for the BookMapper, which stopped NullPointerExceptions occurring for this test and two others.
Problem: testFindBookByNonExistentId() was getting a NullPointerException.
Root Cause: There was no BookMapper mocked for the test. So when BookService.getBookById tried to call bookMapper::toDto, the Exception was being thrown.
Solution: Added a mock for the BookMapper at class level.
Impact: The test is no longer throwing an exception and is now passing. BookServiceTest.testFindBookById() which was also getting a NullPointerException is no longer getting the error (It is now failing for a different reason). BookServiceTest.testSaveBook() is also no longer getting the NullPointerException, but now getting an error for a potential stubbing problem.