fix(test): BookServiceTest.testSaveBook - Got rid of the strict stubbing argument mismatch.
Problem: There was a strict stubbing argument mismatch between bookRepository.save() in the test method and in BookService.java. This problem became evident after the NullPointerException was fixed previously (see previous commit for reference).
Root Cause: Not all the required stubbing statements had been written in the test. bookMapper.toEntity(dto) had not been mocked, so was returning null instead of a book, which was causing the mismatch.
Solution: Added the missing stubbing statements to mock the behaviour of the saveBook() method.
Impact: The test method is now mocking the behaviour of BookService.saveBook() correctly and there is no longer a mismatch.