fix(test): BookServiceTest.testFindBookById - Added more mocks to the test method.
Problem: assertTrue(found.isPresent()) was expecting true but was getting false. This problem became apparent after the NullPointerException was fixed in a previous commit.
Root Cause: Not all the required mocks had been set up for the test.
Solution: Added a stubbing statement for the bookMapper’s behaviour and initialised a BookDto with a matching Id to the mock Book object already in the test.
Impact: The test is now mocking the behaviour of getBookById() correctly and is now passing.