fix(test): TradeServiceTest.testCreateTrade_Success - Added missing variables and mocks needed to avoid errors being thrown, and to accurately mock the behaviour of a new trade object being created.
Problem: Was getting the runtime exception ‘Book not found or not set’
Root Cause: The createTrade() method being called in the test was calling populateReferenceDataByName(), which was then throwing the exception because the tradeDTO object in the test setup did not have a counterparty or book with a name or ID assigned to it. The set up was also missing a TradeStatus and a TradeLeg, which are problems that became apparent after I fixed the initial exception.
Solution: I created mocks for the bookRepository and counterpartyRepository. I declared a Book, Counterparty, TradeStatus and TradeLeg object at class level. I declared them at this scope as I suspect that some of these variables will be needed by TradeServiceTest.testCashflowGeneration_MonthlySchedule(). I assigned values to the objects in setUp(). And I created stubbing statements in the test method to mimic the behaviour of the relevant methods in TradeService.java.
Impact: The test is now passing. Now, exceptions are being thrown and the mocks are mimicking the behaviour of the program when a trade is created.