fix(test): TradeServiceTest.testAmendTrade_Success - Assigned a value to trade.version and added a stubbing statement to get rid of the NullPointerExceptions.
Problem: Test was throwing a NullPointerException, Cannot invoke ‘Java.lang.Integer.intValue()’ because the return value of ‘com.technicalchallenge.model.Trade.getVersion()’ is null
Root Cause: The Trade object in the setup didn’t have a version assigned to it so amendedTrade.setVersion(existingTrade.getVersion() + 1) was throwing a NullPointerException. Additionally, there was no stubbing statement for tradeLegRepository.save(tradeLeg), so leg was null and TradeLeg.getLegId() could not be invoked. This became apparent after I solved the first issue.
Solution: In the test method, I assigned a version integer to the trade object, and I added a stubbing statement for saving the tradeLeg.
Impact: The test is now passing. The NullPointer exceptions are no longer occurring.