fix(test): TradeServiceTest.testCashflowGeneration_MonthlySchedule - Added a parameterised test which calls generateCashflows() indirectly via createTrade().
Problem: This test was a deliberate bug for candidates to find and fix. Candidates were supposed to implement proper cashflow testing
Root Cause: This test method was incomplete and had logical errors. The assertion was written in a way that would always fail.
Solution: I created a schedule with the string “1M” and assigned it to the tradeLeg. I made the test parameterised in order to test the method’s effectiveness with multiple dates. I added stubbing statements to mock the behaviour needed for the generateCashflows() method to execute. generateCashflows() is a private method, so has to be called indirectly through createTrade() or amendTrade(). In this case I used createTrade. And then for each test case, I verified that cashflowRepository.save() was invoked the correct number of times.
Impact: The method generateCashflows() is now being tested correctly